Allow local bot API media above 20MB
