+++ scripts/media-coverage-proof.js
@@ -7,6 +7,7 @@ const {
   transcribableExtension,
   transcribableMediaFromMessage,
 } = require('../dist/helpers/transcribableTelegramMedia')
+const { isMediaTooLarge } = require('../dist/helpers/mediaSizeLimit')
 const { extensionForSource } = require('../dist/workerClient/runWindowsWorker')
 
 function assert(condition, message) {
@@ -110,6 +111,26 @@ assertMedia(
   'document',
   'missing-mime-audio-file'
 )
+
+delete process.env.VOICY_MAX_MEDIA_FILE_SIZE_MB
+assert(
+  !isMediaTooLarge(21 * 1024 * 1024),
+  'default backend media limit should allow files larger than cloud Bot API 20 MB'
+)
+assert(
+  isMediaTooLarge(2048 * 1024 * 1024),
+  'default backend media limit should still cap files at 2048 MB'
+)
+process.env.VOICY_MAX_MEDIA_FILE_SIZE_MB = '20'
+assert(
+  !isMediaTooLarge(20 * 1024 * 1024 - 1),
+  'explicit 20 MB backend media limit should allow files below 20 MB'
+)
+assert(
+  isMediaTooLarge(20 * 1024 * 1024),
+  'explicit 20 MB backend media limit should preserve cloud Bot API behavior'
+)
+delete process.env.VOICY_MAX_MEDIA_FILE_SIZE_MB
 
 assert(
   extensionForSource({
